move and rename the filter class to comply with M4.5 changes.
see MDL-82427 for details.
